Added Evaluate.java to use for getting a utility value
TODO:

When building the tree and searching downwards, when a node is done and returns the utility value should also be return and added to its parent node utility value.

Outside later on the rootnode's children will be checked against each other and there determine which is the best possible move.

Evaluate Class:
In this class the moves, or current moves, will be evaluated to determine what utility value the node should have.